Fixed `AttributeError: portal_windowZ` in show_window template,
Note that this template seems unused and does not actually show
anything when I try to use it, but I may be using it wrongly.

Note that in version 1.3b3 the id `portal_windowZ` was renamed to
`portal_windowz`, with a lowercase `z`.  If you get this
AttributeError when viewing windowz (also with other templates than
show_window) then you have probably upgraded from an old version and
the tool on your website still has the old `portal_windowZ` id.  In
that case you should deactivate the windowZ product in the Add-ons
control panel and activate it again.